api_agents/pyproject.toml
[project]
name = "api-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "langchain>=1.2.15",
    "langchain-anthropic>=1.4.1",
    "langgraph>=1.1.8",
    "pathlib>=1.0.1",
    "pydantic>=2.13.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.44.0",
]
